[UPDATE] 0:31 17/05/2020
Añadidos pop up en la pagina main y ajustado visualización de las dietas con el css.
