+++ Fitness App/FitnessApplication/src/app/continueroutine/continueroutine.page.html
@@ -3,7 +3,7 @@
   <ion-buttons slot="start">
       <ion-back-button></ion-back-button>
     </ion-buttons>
-    <ion-title>Prueba: {{popUpInteract}}</ion-title>
+    <ion-title>Prueba: </ion-title>
   </ion-toolbar>
 </ion-header>
