+++ Fitness App/FitnessApplication/src/app/main/main.page.html
@@ -47,7 +47,7 @@
 
       <ion-content class="body-app">
         <!-- Carrusel de continuar tu rutina -->
-        <ion-item routerLink="/continueroutine">
+        <ion-item (click)="alert()">
           <ion-label>Continuar rutina</ion-label>
           <ion-icon name="arrow-forward-outline" slot="end"></ion-icon>
         </ion-item>
