+++ Fitness App/FitnessApplication/src/app/onereceta/onereceta.page.html
@@ -33,7 +33,7 @@
  <!-- Listado de los alimentos que implementara la receta. -->
   <ion-item *ngFor="let receta of alimentosData" color="transparent">
     <ion-text color="black">
-      <p click={}>{{receta}}</p>
+      <p>{{receta}}</p>
     </ion-text>
     <!--<ion-input placeholder="{{receta}}">{{info[i]}}:  </ion-input>-->
   </ion-item>
