+++ Fitness App/FitnessApplication/src/app/onereceta/onereceta.page.scss
@@ -3,3 +3,15 @@ ion-content {
     background: linear-gradient(180deg, rgb(38, 94, 167) 0%, rgba(73,180,181,1) 100%);
     color: white;
 }
+
+.img-box {
+    padding: 20px 20px 20px 20px;
+}
+
+h1 {
+    margin-left: 10px;
+}
+
+p {
+    margin-left: 10px;
+}
